optimization: shuffle as util function
