+++ src/components/FavoritesView/FavoritesList.svelte
@@ -6,6 +6,7 @@
     playNextList,
     currentID,
     favoritesPlayStatus,
+    shuffle,
   } from '$lib/player';
   import { fade } from 'svelte/transition';
 
@@ -42,14 +43,6 @@
       // if it's not, add it
       $playNextList = [...$playNextList, result];
     });
-  }
-  function shuffle(array: FavoriteStore[]) {
-    const newArray = [...array];
-    for (let i = newArray.length - 1; i > 0; i--) {
-      const j = Math.floor(Math.random() * (i + 1));
-      [newArray[i], newArray[j]] = [newArray[j], newArray[i]];
-    }
-    return newArray;
   }
 
   async function togglePlayAll() {
