+++ src/lib/player.ts
@@ -60,6 +60,15 @@ export function secondsToTime(seconds: number) {
     return h + ':' + m + ':' + s;
 };
 
+export function shuffle<T>(array: T[]): T[] {
+    const newArray = [...array];
+    for (let i = newArray.length - 1; i > 0; i--) {
+        const j = Math.floor(Math.random() * (i + 1));
+        [newArray[i], newArray[j]] = [newArray[j], newArray[i]];
+    }
+    return newArray;
+}
+
 
 // ------------------------------------------------------------
 // HASH Management
